More thoughts on paper development
I think Immerzeel et al. (2019) provides an excellent baseline to develop this paper. We can re-use many of the indices used worldwide, and present AIRs as a sustainable solution for regions with high risk. I still have not checked the spatial resolution of those indices, but I have also added additional datasets which we could use to run more high-resolution analysis if we found that the products from Immerzeel et al. (2019) are not detailed enough.